Consolidate chat-template token-id extraction into a shared helper
Move get_conversation_input_ids into modelopt.torch.speculative.utils as a single source
of truth and call it from all three offline-dump entry points that shared the BatchEncoding
bug: compute_hidden_states_trtllm.py and the send_conversation{,_vllm} scripts (which also
indexed/decoded the BatchEncoding). Drop the now-dead torch.Tensor branch. The helper is
torch-only, so the test imports it directly via get_tiny_tokenizer -- revert the
speculative_decoding lane back to the torch container.
