+++ .github/workflows/example_tests.yml
@@ -61,12 +61,13 @@ jobs:
               - tests/examples/conftest.py
               - "!**/*.{md,rst,png,ipynb}"
             torch:
-              - "{examples,tests/examples}/{llm_distill,llm_qat,llm_sparsity,specdec_bench}/**"
-              - examples/dataset/** # data prep for llm_qat
+              - "{examples,tests/examples}/{llm_distill,llm_qat,llm_sparsity,specdec_bench,speculative_decoding}/**"
+              - examples/dataset/** # data prep for llm_qat and speculative_decoding
+              - examples/hf_ptq/** # the speculative_decoding test drives hf_ptq
               - "!**/*.{md,rst,png,ipynb}"
             trtllm:
-              - "{examples,tests/examples}/{gpt-oss,hf_ptq,llm_eval,speculative_decoding}/**"
-              - examples/dataset/** # data prep for gpt-oss and speculative_decoding
+              - "{examples,tests/examples}/{gpt-oss,hf_ptq,llm_eval}/**"
+              - examples/dataset/** # data prep for gpt-oss
               - "!**/*.{md,rst,png,ipynb}"
             megatron:
               - "{examples,tests/examples}/megatron_bridge/**"
@@ -111,7 +112,7 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
-        example: [llm_distill, llm_qat, llm_sparsity, specdec_bench]
+        example: [llm_distill, llm_qat, llm_sparsity, specdec_bench, speculative_decoding]
     uses: ./.github/workflows/_example_tests_runner.yml
     permissions:
       contents: read
@@ -131,7 +132,7 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
-        example: [gpt-oss, hf_ptq, llm_eval, speculative_decoding]
+        example: [gpt-oss, hf_ptq, llm_eval]
     uses: ./.github/workflows/_example_tests_runner.yml
     permissions:
       contents: read
